refactor(gmail): extract applyRejection service for bot reuse (#456)
